Experimental GetTextContentAsync
GetTextContentSync is slow - largely due to the serialization and then deserialization of a large amount of data when passing from CefSharp to the PageRenderer.

Attempt to counteract this with the Async code - the idea is we should be able to request more text content from CefSharp while deserializing what we've already got.

Profiling seems to show a doubling (at best) in speed.