Added blocking PageRenderer constructor
Prior to this commit, the PageRenderer object could only be created via the Async 'Create' static method.

To ease confusion, 'Create' has been re-labelled 'CreateAsync', and a standard PageRenderer() constructor has been created that will block the calling thread until the PageRenderer is set up and running.

Note this may block for some time if Chrome needs to be downloaded to the machine, so in many use cases it may not be desirable to block the thread in this way and 'CreateAsync' might be more appropriate.