Minor loading fix
After the previous update, we've streamlined all of the JavaScript functions that fire through Puppeteer to return a promise. In doing so, we're not waiting on additional callback events.

There is a _pdfLoadWaiting flag that prevented two PDFs from loading at the same time, but since the callbacks have been removed, it was never getting set to 'false' - meaning further PDF file loads failed after the first load.