Add ability to set maximum width and height when rendering pages
The pixel size of a PDF page is directly related to the page dimensions - therefore, when rendering with a page scale, the resulting image would not be limited to any particular size.

By using RenderPage with a maxWidth and maxHeight setting, it's possible to ensure that the page image is not going to be ridiculously huge.