Allow nullable types from PDF.JS
I found a PDF with a font style that had null ascent and descent values. Since PDF2IMG is expecting doubles, these are non-nullable and it fired an exception. Sorted now to make the doubles nullable from JavaScript. We then later convert them to non-nullable values and decide what to do with nulls at that point. This is used in the 'GetTextContent' call.